+++ src/day8/re_frame_10x.cljs
@@ -151,6 +151,7 @@
 (goog-define time-travel? true)
 (goog-define ignored-libs "[:reagent :re-frame]")
 (goog-define ns-aliases "{long-namespace ln}")
+(goog-define trace-when ":panel")
 
 (def project-config
   (let [read      reader.edn/read-string-maybe
@@ -163,5 +164,6 @@
           :filtered-view-trace    (some->> hidden-namespaces read (map view) sortable-uuid-map)
           :app-db-follows-events? time-travel?
           :low-level-trace        (some-> ignored-libs read (pred-map #{:re-frame :reagent}))
-          :ns-aliases             (some->> ns-aliases read (map alias) sortable-uuid-map)}
+          :ns-aliases             (some->> ns-aliases read (map alias) sortable-uuid-map)
+          :trace-when             (some->> trace-when read keyword)}
          (into {} keep-vals))))