+++ src/day8/re_frame_10x/panels/settings/subs.cljs
@@ -188,3 +188,8 @@
  ::expansion-limit
  :<- [::root]
  :-> :expansion-limit)
+
+(rf/reg-sub
+ ::trace-when
+ :<- [::root]
+ :-> :trace-when)